requirements.txt
jsonpickle==2.0.0
matplotlib==3.4.3
antlr4-python3-runtime==4.7.2
numpy==1.21.2
scipy==1.8.0
simplejson==3.16.0
importlib==1.0.4
requests==2.22.0
gitdb==4.0.5
gitpy==0.0.5
GitPython==3.1.14